change update function
