+++ src/plugin/pluginData.ts
@@ -1,21 +1,13 @@
 import omit from 'just-omit';
 import get from 'just-safe-get';
 import { isEqual } from '@/utils/isEqual';
-import { NodeTokenRefMap } from '@/types/NodeTokenRefMap';
-import { SharedPluginDataNamespaces } from '@/constants/SharedPluginDataNamespaces';
 import { Properties } from '@/constants/Properties';
-import { MessageFromPluginTypes } from '@/types/messages';
-import { BackgroundJobs } from '@/constants/BackgroundJobs';
-import { AnyTokenList } from '@/types/tokens';
 import store from './store';
-import { notifySelection, postToUI } from './notifiers';
+import { notifySelection } from './notifiers';
 import removeValuesFromNode from './removeValuesFromNode';
 import { defaultNodeManager, NodeManagerNode } from './NodeManager';
 import { tokensSharedDataHandler } from './SharedDataHandler';
-import { defaultWorker } from './Worker';
-import { ProgressTracker } from './ProgressTracker';
 import { SelectionGroup, SelectionValue } from '@/types';
-import { CompositionTokenProperty } from '@/types/CompositionTokenProperty';
 import { TokenTypes } from '@/constants/TokenTypes';
 import getAppliedStylesFromNode from './getAppliedStylesFromNode';
 import getAppliedVariablesFromNode from './getAppliedVariablesFromNode';
@@ -163,110 +155,4 @@ export async function setNonePluginData({ nodes, key }: { nodes: readonly (BaseN
     removeValuesFromNode(node, key);
     store.successfulNodes.add(node);
   }));
-}
-
-export async function updatePluginData({
-  entries, values, shouldOverride = false, shouldRemove = true, tokensMap,
-}: { entries: readonly NodeManagerNode[], values: NodeTokenRefMap, shouldOverride?: boolean, shouldRemove?: boolean, tokensMap?: Map<string, AnyTokenList[number]> }) {
-  // Big O (n * m): (n = amount of nodes, m = amount of applied tokens to the node)
-  const namespace = SharedPluginDataNamespaces.TOKENS;
-  postToUI({
-    type: MessageFromPluginTypes.START_JOB,
-    job: {
-      name: BackgroundJobs.PLUGIN_UPDATEPLUGINDATA,
-      timePerTask: 2,
-      completedTasks: 0,
-      totalTasks: entries.length,
-    },
-  });
-
-  const tracker = new ProgressTracker(BackgroundJobs.PLUGIN_UPDATEPLUGINDATA);
-  const promises: Set<Promise<void>> = new Set();
-  entries.forEach(({ node, tokens }) => {
-    promises.add(defaultWorker.schedule(async () => {
-      const figmaNode = figma.getNodeById(node.id);
-      if (!figmaNode) {
-        console.warn(`Node with ID ${node.id} no longer exists`);
-        return;
-      }
-      const currentValuesOnNode = tokens ?? {};
-      let newValuesOnNode: NodeTokenRefMap = {};
-      if (values.composition === 'delete') newValuesOnNode = { ...values, ...currentValuesOnNode, composition: values.composition };
-      else newValuesOnNode = { ...currentValuesOnNode, ...values };
-      if (currentValuesOnNode.composition && values.composition) {
-        // when select another composition token, reset applied properties by current composition token
-        const resolvedToken = tokensMap?.get(currentValuesOnNode.composition);
-        let removeProperties: string[] = [];
-        if (resolvedToken && resolvedToken.rawValue) {
-          removeProperties = Object.keys(resolvedToken.rawValue).map((property) => (
-            property
-          ));
-        }
-        if (removeProperties && removeProperties.length > 0) {
-          await Promise.all(removeProperties.map(async (property) => {
-            await removePluginData({ nodes: [node], key: property as Properties, shouldRemoveValues: shouldRemove });
-          }));
-        }
-        shouldOverride = true;
-      }
-      await Promise.all(Object.entries(newValuesOnNode).map(async ([key, value]) => {
-        if (value === currentValuesOnNode[key as CompositionTokenProperty] && !shouldOverride) {
-          return;
-        }
-
-        const jsonValue = JSON.stringify(value);
-        switch (value) {
-          case 'delete':
-            delete newValuesOnNode[key as CompositionTokenProperty];
-            await removePluginData({ nodes: [node], key: key as Properties, shouldRemoveValues: shouldRemove });
-            break;
-          case 'none':
-            await setNonePluginData({ nodes: [node], key: key as Properties });
-            break;
-          // Pre-Version 53 had horizontalPadding and verticalPadding.
-          case 'horizontalPadding':
-            newValuesOnNode.paddingLeft = jsonValue;
-            newValuesOnNode.paddingRight = jsonValue;
-            node.setSharedPluginData(namespace, Properties.paddingLeft, jsonValue);
-            node.setSharedPluginData(namespace, Properties.paddingRight, jsonValue);
-            break;
-          case 'verticalPadding':
-            newValuesOnNode.paddingTop = jsonValue;
-            newValuesOnNode.paddingBottom = jsonValue;
-            node.setSharedPluginData(namespace, Properties.paddingTop, jsonValue);
-            node.setSharedPluginData(namespace, Properties.paddingBottom, jsonValue);
-            break;
-          default:
-            node.setSharedPluginData(namespace, key, jsonValue);
-            break;
-        }
-      }));
-      await defaultNodeManager.updateNode(node, newValuesOnNode);
-
-      const nodeHasNoValues = Object.keys(newValuesOnNode).length === 0 && newValuesOnNode.constructor === Object;
-      const editRelaunchData = node.getRelaunchData() as {
-        edit?: string
-      };
-
-      if (!nodeHasNoValues) {
-        const updatedRelaunchData = Object.keys(newValuesOnNode).join(', ');
-        if (updatedRelaunchData !== editRelaunchData.edit) {
-          node.setRelaunchData({
-            edit: updatedRelaunchData,
-          });
-        }
-      } else {
-        node.setRelaunchData({});
-      }
-
-      tracker.next();
-      tracker.reportIfNecessary();
-    }));
-  });
-  await Promise.all(promises);
-
-  postToUI({
-    type: MessageFromPluginTypes.COMPLETE_JOB,
-    name: BackgroundJobs.PLUGIN_UPDATEPLUGINDATA,
-  });
 }
